feat: DiaryService에 인증 정보 확인 로직 추가. 현재 인증된 사용자의 정보를 반환하는 유틸리티 클래스 AuthenticationUtils 추가. 그 외 로직 구현을 위한 DTO 수정과 메소드 추가.

+++ src/main/java/org/diarymoodanalyzer/service/DiaryService.java
@@ -15,11 +15,16 @@ import org.diarymoodanalyzer.dto.response.GetDiaryByPageResponse;
 import org.diarymoodanalyzer.dto.response.GetDiaryTitleByPageResponse;
 import org.diarymoodanalyzer.repository.DiaryRepository;
 import org.diarymoodanalyzer.repository.UserRepository;
+import org.diarymoodanalyzer.util.AuthenticationUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @RequiredArgsConstructor
 @Service
@@ -34,18 +39,21 @@ public class DiaryService {
 
     //트랜잭션으로 관리
     @Transactional
-    public AddDiaryResponse addDiary(AddDiaryRequest dto) {
+    public AddDiaryResponse addDiary(AddDiaryRequest dto) throws ResponseStatusException {
 
-        /*
-        실제로는 토큰이 해당 유저에 대해 유효한지 체크하는 과정도 필요함.
-         */
+        //현재 인증된 사용자의 정보를 가져옴
+        String currentUserEmail = AuthenticationUtils.getCurrentUserEmail();
+
+        if(currentUserEmail == null) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "There is no Authentication");
+        }
 
         //이메일을 통해 User의 아이디 값만 받아온다.
-        Long userId = userRepository.findIdByEmail(dto.getEmail());
+        Long userId = userRepository.findIdByEmail(currentUserEmail);
 
         //못 찾으면 예외
         if(userId == null) {
-            throw new EntityNotFoundException("not found user with email : " + dto.getEmail());
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "not found user with email : " + currentUserEmail);
         }
 
         //프록시 객체로 받아온다. 실제로 엔티티에 접근하기 전에는 DB에 쿼리하지 않는다.
@@ -61,7 +69,7 @@ public class DiaryService {
         //리포지토리에 저장
         Diary savedDiary = diaryRepository.save(diary);
 
-        return new AddDiaryResponse(savedDiary.getDiaryId(), dto.getEmail(), dto.getTitle());
+        return new AddDiaryResponse(savedDiary.getDiaryId(), currentUserEmail, dto.getTitle());
 
         /*
         프록시 객체의 속성에 접근하지 않고 단순히 diary에 참조만 넘겼다.
@@ -70,9 +78,22 @@ public class DiaryService {
          */
     }
 
-    public GetDiaryByIdResponse getDiaryById(long id) {
+    public GetDiaryByIdResponse getDiaryById(long id) throws ResponseStatusException {
+
+        //diaryId로 소유자 userId 조회
+        Long ownerId = diaryRepository.findUserIdById(id);
+
+        if(ownerId == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "not found with diary id : " + id);
+        }
+
+        //현재 인증된 유저의 id가 소유자의 id와 일치하지 않는다면 예외
+        if(isInvalidAuthByUserId(ownerId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You have not permission");
+        }
+
         return new GetDiaryByIdResponse(diaryRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("not found with userId : " + id)
+                () -> new IllegalArgumentException("not found with diary id : " + id)
         ));
     }
 
@@ -84,7 +105,10 @@ public class DiaryService {
         //페이지의 번호와 사이즈, 정렬 기준을 지정
         Pageable pageable = PageRequest.of(req.getPage(), req.getSize(), sortBy);
 
-        Page<Diary> diaries = diaryRepository.findByUserEmail(req.getEmail(), pageable);
+        //현재 인증된 유저 정보 불러옴
+        String currentUserEmail = AuthenticationUtils.getCurrentUserEmail();
+
+        Page<Diary> diaries = diaryRepository.findByUserEmail(currentUserEmail, pageable);
 
         //리포지토리를 통해 페이지로 받아서 컨트롤러로 반환
         //DTO의 생성자로 매핑하여, 원본 엔티티가 아닌 DTO의 페이지로 리턴한다.
@@ -98,12 +122,49 @@ public class DiaryService {
         //페이지의 번호와 사이즈, 정렬 기준을 지정
         Pageable pageable = PageRequest.of(req.getPage(), req.getSize(), sortBy);
 
-        return diaryRepository.findByUserEmailOnlyTitle(req.getEmail(), pageable);
+        //현재 인증된 유저 정보 불러옴
+        String currentUserEmail = AuthenticationUtils.getCurrentUserEmail();
+
+        return diaryRepository.findByUserEmailOnlyTitle(currentUserEmail, pageable);
     }
 
-    public void deleteDiaryById(long id) {
+    public void deleteDiaryById(long id) throws ResponseStatusException {
+
+        Long ownerId = diaryRepository.findUserIdById(id);
+
+        if(ownerId == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "not found with diary id : " + id);
+        }
+
+        if(isInvalidAuthByUserId(ownerId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You have not permission");
+        }
+
         diaryRepository.deleteById(id);
     }
 
+    /*
+    현재 인증된 유저의 이메일이, 인자로 받은 이메일과 같으면 true.
+    otherwise, false
+     */
+    private boolean validateAuthByEmail(String email) {
+        String currentUserEmail = AuthenticationUtils.getCurrentUserEmail();
+        return currentUserEmail != null && currentUserEmail.equals(email);
+    }
 
+    /*
+    현재 인증된 유저의 아이디 값이, 인자로 받은 아이디값과 같으면 true.
+    otherwise, false
+     */
+    private boolean validateAuthByUserId(Long id) {
+        Long userId = AuthenticationUtils.getCurrentUserId();
+        return userId != null && userId.equals(id);
+    }
+
+    /*
+    inverted method of validateAuthByUserId(Long)
+     */
+    private boolean isInvalidAuthByUserId(Long id) {
+        return !validateAuthByUserId(id);
+    }
 }
